x2文件系统移植完成，正在设计二级加载器 ---------- 花了两天时间，完成了仅仅依靠接口的文件系统 这将使该文件系统具备完整的、100%的可移植性，并且也指明了将来对于内核其他部分的设计思想。 将BIOS看成0级加载器，MBR(VBR)看成1级加载器，将类grub的引导程序看成2级加载程序，从而实现内核的加载分层标准。目前已经完成了MBR的设计。

+++ include/macros/AssociatedMemoryManager_macros.h
@@ -88,7 +88,7 @@ size_t AssociatedMemoryManager<T, 1>::getTargetIndex(TargetType* t)const
 {
 //	ptrdiff_t diff=((char*)t - (char*)this->tarr)/sizeof(*t);
 	ptrdiff_t diff=t - this->tarr;
-	return (t==NULL || diff<0||diff >= this->len) ?(this->len +1 ):(size_t)diff;
+	return (t==NULL || diff<0||(size_t)diff >= this->len) ?(this->len +1 ):(size_t)diff;
 
 }
 template<class T>
@@ -96,7 +96,7 @@ size_t		AssociatedMemoryManager<T, 1>::getNodeIndex(NodeType* n)const
 {
 //	ptrdiff_t diff=((char*)n - (char*)this->narr)/sizeof(*n);
 	ptrdiff_t diff=n - this->narr;
-	return (n==NULL||diff<0||diff >= this->len) ?(this->len +1 ):(size_t)diff;
+	return (n==NULL||diff<0||(size_t)diff >= this->len) ?(this->len +1 ):(size_t)diff;
 }
 template<class T>
 size_t		AssociatedMemoryManager<T, 1>::getEachSize()
